Extract SMPTE utilities into own package
