+++ .changeset/ten-banks-live.md
@@ -1,0 +1,8 @@
+---
+'@arcanewizards/smpte': minor
+---
+
+Initial implementation of package
+
+Extract common utilities and constants from the `@arcanewizards/artnet` package
+so it can be used with other SMPTE-based protocols (namely MIDI)
