+++ packages/artnet/src/constants.ts
@@ -1,17 +1,20 @@
+import { SMPTE_TIMECODE_FPS, SMPTETimecodeMode } from '@arcanewizards/smpte';
+
 export const ARTNET_PORT = 6454;
 
-export const TIMECODE_MODES = {
+export const TIMECODE_MODES: Record<SMPTETimecodeMode, number> = {
   FILM: 0,
   EBU: 1,
   DF: 2,
   SMPTE: 3,
 };
 
-export type TimecodeMode = keyof typeof TIMECODE_MODES;
+/**
+ * @deprecated Use `SMPTE_TIMECODE_FPS` from `@arcanewizards/smpte` instead.
+ */
+export const TIMECODE_FPS = SMPTE_TIMECODE_FPS;
 
-export const TIMECODE_FPS: Record<TimecodeMode, number> = {
-  FILM: 24,
-  EBU: 25,
-  DF: 29.97,
-  SMPTE: 30,
-};
+/**
+ * @deprecated Use `TimecodeMode` from `@arcanewizards/smpte` instead.
+ */
+export type TimecodeMode = SMPTETimecodeMode;
