+++ packages/smpte/README.md
@@ -1,0 +1,11 @@
+# `@arcanewizards/midi-timecode`
+
+[![](https://img.shields.io/npm/v/@arcanewizards/smpte)](https://www.npmjs.com/package/@arcanewizards/smpte)
+
+Small library for handling processing & calculations of SMPTE timecodes.
+
+## Installation
+
+```sh
+pnpm add @arcanewizards/smpte
+```
