+++ packages/smpte/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "@arcanewizards/typescript-config/base.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "rootDir": "src",
+    "moduleResolution": "nodenext",
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"],
+}
